Weitere Aktionen gedebuggt Gasplaneten werden beim ersten Aufstarten Initialisiert (1. Update zur Loginzeit)
Nachtrag vom letzten Update:
Wichtigste Änderung, dass beim Rückflug überprüft wird, ob der planet_user_id=fleet.user_id und die Flotte sonst zum Main gesendet wird.

+++ fleethandler/fleetActions/debris/DebrisHandler.cpp
@@ -76,7 +76,7 @@ namespace debris
 						{
 							rRow = rRes.at(i);
 
-							this->cnt = (int)ceil((int)rRow["fs_ship_cnt"]*0.4);
+							this->cnt = ceil((double)rRow["fs_ship_cnt"]*0.4);
 							this->tfMetal += cnt * (double)rRow["ship_costs_metal"];
 							this->tfCrystal += cnt * (double)rRow["ship_costs_crystal"];
 							this->tfPlastic += cnt * (double)rRow["ship_costs_plastic"];
